reorganize folder
make the files more organize.
still there are outstanding errors on the swipe list demo
