+++ UltimateRecyclerView/app/src/main/AndroidManifest.xml
@@ -28,34 +28,34 @@
         </activity>
 
         <activity
-            android:name=".MultiViewTypesActivity"
+            android:name=".multiitemdemo.MultiViewTypesActivity"
             android:label="MultiViewTypesActivity" />
         <activity
-            android:name=".GridLayoutRVTest"
+            android:name=".griddemo.GridLayoutRVTest"
             android:label="GridLayoutRVTest" />
         <activity
-            android:name=".CustomSwipeToRefreshRefreshActivity"
+            android:name=".loadmoredemo.CustomSwipeToRefreshRefreshActivity"
             android:label="CustomSwipeToRefreshRefreshActivity" />
         <activity
-            android:name=".TestAdMob"
+            android:name=".admobdemo.TestAdMob"
             android:label="TestAdMob" />
         <activity
-            android:name=".TestExpandableRV"
+            android:name=".expandemo.TestExpandableRV"
             android:label="TestExpandableRV" />
         <activity
-            android:name=".DebugLoadMoreActivity"
+            android:name=".loadmoredemo.DebugLoadMoreActivity"
             android:label="DebugLoadMoreActivity" />
         <activity
             android:name=".scrollableobservable.ScrollObservablesActivity"
             android:label="ScrollObservablesActivity" />
         <activity
-            android:name=".DragActivity"
+            android:name=".dragdemo.DragActivity"
             android:label="DragActivity" />
         <activity
-            android:name=".swipelist.SwipeListViewExampleActivity"
+            android:name=".swiplistdemo.SwipeListViewExampleActivity"
             android:label="SwipeListViewExampleActivity" />
         <activity
-            android:name=".TestAdvancedAdmobActivity"
+            android:name=".admobdemo.TestAdvancedAdmobActivity"
             android:label="TestAdvancedAdmobActivity" />
 
         <activity
