+++ UltimateRecyclerView/app/src/main/java/com/marshalchen/ultimaterecyclerview/demo/basicdemo/SimpleAnimationAdapter.java
@@ -1,4 +1,4 @@
-package com.marshalchen.ultimaterecyclerview.demo;
+package com.marshalchen.ultimaterecyclerview.demo.basicdemo;
 
 import android.animation.Animator;
 import android.graphics.Color;
@@ -15,6 +15,7 @@ import com.marshalchen.ultimaterecyclerview.URLogs;
 import com.marshalchen.ultimaterecyclerview.UltimateRecyclerviewViewHolder;
 import com.marshalchen.ultimaterecyclerview.UltimateViewAdapter;
 import com.marshalchen.ultimaterecyclerview.animators.internal.ViewHelper;
+import com.marshalchen.ultimaterecyclerview.demo.R;
 
 import java.security.SecureRandom;
 import java.util.List;
