+++ UltimateRecyclerView/app/src/main/res/layout/activity_drag.xml
@@ -1,6 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
@@ -35,8 +36,8 @@
                 android:id="@+id/add"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_toLeftOf="@id/del"
                 android:layout_centerInParent="true"
+                android:layout_toLeftOf="@id/del"
                 android:background="?attr/selectableItemBackground"
                 android:padding="10dp"
                 android:text="Add" />
@@ -45,12 +46,12 @@
     </android.support.v7.widget.Toolbar>
 
     <android.support.v7.widget.RecyclerView
+        android:id="@+id/ultimate_recycler_view"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
-        android:id="@+id/ultimate_recycler_view"
+        android:layout_marginTop="?attr/actionBarSize"
         app:recyclerviewClipToPadding="true"
-        app:recyclerviewFloatingActionView="@layout/floating_view"
-        app:recyclerviewScrollbars="vertical"
         app:recyclerviewDefaultSwipeColor="@array/google_colors"
-        android:layout_marginTop="?attr/actionBarSize" />
+        app:recyclerviewFloatingActionView="@layout/floating_view"
+        app:recyclerviewScrollbars="vertical" />
 </FrameLayout>
